requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Machine Learning
scikit-learn>=1.2.0

# NLP
nltk>=3.8.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
wordcloud>=1.8.0
plotly>=5.15.0

# Web Application
streamlit>=1.28.0

# Deep Learning
tensorflow>=2.13.0
transformers>=4.30.0
torch>=2.0.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
PyYAML>=6.0.0

# Development
pytest>=7.0.0
